pyproject.toml
[project]
name = "chippypi"
version = "3.0.0"
description = "KOSMOS Chipz hexapod robot — custom controller"
requires-python = ">=3.11"
dependencies = [
    "pigpio",
    "valkey[libvalkey]",
    "acconeer-exptool[a121]",
    "nicegui",
    "loguru",
    "orjson",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"